修复: SD_Verify 移到 FreeRTOS 任务 + PA8 卡检测 GPIO
1. SD_Verify() 从 main() 移到 initTestTask（FreeRTOS 任务），修复调用时序错误（sd_diskio 依赖 RTOS 消息队列，需在调度器启动后运行）
2. 加入 PA8 物理卡检测：新增 SD_Detect_GPIO_Init()，BSP_SD_IsDetected() 改为读取 PA8 电平（内部上拉，LOW=卡存在）
3. 修复 sd_verify.c printf %lu 格式警告

+++ CANbus_code/Core/Src/sd_verify.c
@@ -35,7 +35,7 @@ uint8_t SD_Verify(void)
     sprintf(buf, "[SD] FAIL: f_mount error %d\r\n", fr);
     sd_print(buf);
     sprintf(buf, "[SD] HAL SD state=%lu error=%lu\r\n",
-            hsd1.State, hsd1.ErrorCode);
+            (unsigned long)hsd1.State, (unsigned long)hsd1.ErrorCode);
     sd_print(buf);
     return 0;
   }
